updates to controllers and requests to process tokens

+++ src/backend/Lifelog/Peace.Lifelog.MediaMementoWebService/Controllers/MediaMementoController.cs
@@ -6,6 +6,8 @@ using System;
 using System.Text.Json;
 using System.Threading.Tasks;
 
+using back_end;
+using Peace.Lifelog.Security;
 using DomainModels;
 using Peace.Lifelog.Logging;
 using Peace.Lifelog.MediaMementoService;
@@ -16,25 +18,92 @@ public class MediaMementoController : ControllerBase
 {
     private readonly IMediaMementoService _mediaMementoService;
     private readonly ILogging _logger;
+    private JWTService jwtService;
 
     public MediaMementoController(IMediaMementoService mediaMementoService, ILogging logger)
     {
         _mediaMementoService = mediaMementoService;
+        this.jwtService = new JWTService();
         _logger = logger;
     }
 
     [HttpPost("UploadMedia")]
     public async Task<IActionResult> UploadMediaMemento([FromBody] UploadMediaMementoRequest payload)
     {
-        var response = await _mediaMementoService.UploadMediaMemento(payload.UserHash, payload.LliId, payload.Binary);
-        return Ok(JsonSerializer.Serialize<Response>(response));
+        try
+        {
+            if (Request.Headers == null)
+            {
+                return StatusCode(401);
+            }
+
+            var jwtToken = JsonSerializer.Deserialize<Jwt>(Request.Headers["Token"]!);
+
+            if (jwtToken == null)
+            {
+                return StatusCode(401);
+            }
+
+            var userHash = jwtToken.Payload.UserHash;
+
+            if (userHash == null)
+            {
+                return StatusCode(401);
+            }
+
+            if (!jwtService.IsJwtValid(jwtToken))
+            {
+                return StatusCode(401);
+            }
+
+            var response = await _mediaMementoService.UploadMediaMemento(userHash, payload.LliId, payload.Binary);
+    
+            return Ok(response);
+        }
+        catch (Exception ex)
+        {
+            _ = await _logger.CreateLog("Logs", "MapsController", "ERROR", "System", ex.Message);
+            return StatusCode(500, "An error occurred while processing your request.");
+        }
     }
 
     [HttpPost("DeleteMedia")]
     public async Task<IActionResult> UploadMediaMemento([FromBody] DeleteMediaMementoRequest payload)
     {
-        var response = await _mediaMementoService.DeleteMediaMemento(payload.LliId);
-        return Ok(JsonSerializer.Serialize<Response>(response));
+        try
+        {
+            if (Request.Headers == null)
+            {
+                return StatusCode(401);
+            }
+
+            var jwtToken = JsonSerializer.Deserialize<Jwt>(Request.Headers["Token"]!);
+
+            if (jwtToken == null)
+            {
+                return StatusCode(401);
+            }
+
+            var userHash = jwtToken.Payload.UserHash;
+
+            if (userHash == null)
+            {
+                return StatusCode(401);
+            }
+
+            if (!jwtService.IsJwtValid(jwtToken))
+            {
+                return StatusCode(401);
+            }
+
+            var response = await _mediaMementoService.DeleteMediaMemento(payload.LliId);
+            return Ok(response);
+        }
+        catch (Exception ex)
+        {
+            _ = await _logger.CreateLog("Logs", "MapsController", "ERROR", "System", ex.Message);
+            return StatusCode(500, "An error occurred while processing your request.");
+        }
     }
 
 }
